Finishing up tests
Moved the helper functions into a helpers package for reuse.
Created a blackbox package and moved the existing tests.
Extracted tests into files containing similar tests to allow the setup
method to be more specific and cut down on repeated code.
