+++ connecttestapp/src/androidTest/java/com/zendesk/connect/testapp/blackbox/AnyExceptionTests.kt
@@ -1,6 +1,9 @@
-package com.zendesk.connect.testapp
+package com.zendesk.connect.testapp.blackbox
 
 import android.support.test.rule.ActivityTestRule
+import com.zendesk.connect.testapp.MainActivity
+import com.zendesk.connect.testapp.helpers.clearDatabase
+import com.zendesk.connect.testapp.helpers.clearSharedPrefs
 import io.outbound.sdk.Event
 import io.outbound.sdk.Outbound
 import io.outbound.sdk.User
@@ -10,16 +13,24 @@ import org.junit.Rule
 import org.junit.Test
 import org.mockito.ArgumentMatchers.anyString
 
+const val configPath = "/i/config/sdk/android"
+const val identifyPath = "/v2/identify"
+const val trackPath = "/v2/track"
+const val registerPath = "/v2/gcm/register"
+const val disablePath = "/v2/gcm/disable"
+const val pairPath = "/i/testsend/push/pair/android"
+
 /**
  * Tests any exceptions expected by using the SDK non-init methods before initialisation.
  *
- * The tests that expect NullPointerException must run first due to poor thread handling in the SDK.
- * Tests which successfully call init will result in the WorkerThread existing leading to an
- * IllegalStateException rather than a NullPointerException. Simple ensuring that this file is first
- * alphabetically will cause the tests to run first.
+ * Even though storage is cleared between tests, the extra threads spun up may not stop
+ * between tests. This means that we will encounter an IllegalStateException instead of a
+ * NullPointerException for any tests that run after tests which successfully call init.
  *
- * The tests for getActiveToken and pairDevice aren't dependent on the WorkerThread but I've
- * included them here to keep them together with the other exception tests.
+ * None of the tests in this file should result in successful initialisations so ensuring that
+ * this file is first alphabetically will cause the tests to run before any initialisations.
+ *
+ * getActiveToken and pairDevice aren't dependent on the WorkerThread so they will throw IllegalStateException.
  */
 class AnyExceptionTests {
 
@@ -28,12 +39,10 @@ class AnyExceptionTests {
 
     @Before
     fun setUp() {
+        clearSharedPrefs()
+        clearDatabase()
+
         testRule.launchActivity(null)
-    }
-
-    @After
-    fun tearDown() {
-
     }
 
     @Test(expected = NullPointerException::class)
